Expanded row tabs change data on other expanded rows
This is due to changing tabs using the same Redux widget. This removes the Redux code in favor of managing tabs locally.

Fixes https://github.com/project-koku/koku-ui/issues/849
Fixes https://github.com/project-koku/koku-ui/issues/848
