+++ src/pages/awsDetails/detailsWidgetView.tsx
@@ -1,4 +1,9 @@
-import { Title } from '@patternfly/react-core';
+import { Button, ButtonType, ButtonVariant } from '@patternfly/react-core';
+import { css } from '@patternfly/react-styles';
+import {
+  Skeleton,
+  SkeletonSize,
+} from '@red-hat-insights/insights-frontend-components/components/Skeleton';
 import { AwsQuery, getQuery } from 'api/awsQuery';
 import { AwsReport, AwsReportType } from 'api/awsReports';
 import {
@@ -10,18 +15,30 @@ import { InjectedTranslateProps, translate } from 'react-i18next';
 import { connect } from 'react-redux';
 import { awsReportsActions, awsReportsSelectors } from 'store/awsReports';
 import { createMapStateToProps, FetchStatus } from 'store/common';
+import { getTestProps, testIds } from 'testIds';
 import { formatValue } from 'utils/formatValue';
-import { formatCurrency } from 'utils/formatValue';
-import { styles } from './detailsWidgetModal.styles';
+import { ComputedAwsReportItem } from 'utils/getComputedAwsReportItems';
+import { getComputedAwsReportItems } from 'utils/getComputedAwsReportItems';
+import { AwsDetailsTab } from './detailsWidget';
+import { styles } from './detailsWidget.styles';
+import { DetailsWidgetModal } from './detailsWidgetModal';
+import { DetailsWidgetModalViewProps } from './detailsWidgetModalView';
 
 interface DetailsWidgetViewOwnProps {
   groupBy: string;
+  item: ComputedAwsReportItem;
+  parentGroupBy: string;
 }
 
 interface DetailsWidgetViewStateProps {
-  queryString?: string;
-  report?: AwsReport;
-  reportFetchStatus?: FetchStatus;
+  availableTabs?: AwsDetailsTab[];
+  queryString: string;
+  report: AwsReport;
+  reportFetchStatus: FetchStatus;
+}
+
+interface DetailsWidgetViewState {
+  isWidgetModalOpen: boolean;
 }
 
 interface DetailsWidgetViewDispatchProps {
@@ -36,55 +53,128 @@ type DetailsWidgetViewProps = DetailsWidgetViewOwnProps &
 const reportType = AwsReportType.cost;
 
 class DetailsWidgetViewBase extends React.Component<DetailsWidgetViewProps> {
-  constructor(props: DetailsWidgetViewProps) {
-    super(props);
-  }
+  public state: DetailsWidgetViewState = {
+    isWidgetModalOpen: false,
+  };
 
   public componentDidMount() {
     const { fetchReport, queryString } = this.props;
     fetchReport(reportType, queryString);
   }
 
-  public componentDidUpdate(prevProps: DetailsWidgetViewProps) {
+  public componentDidUpdate(prevProps: DetailsWidgetModalViewProps) {
     const { fetchReport, queryString } = this.props;
     if (prevProps.queryString !== queryString) {
       fetchReport(reportType, queryString);
     }
   }
 
-  public render() {
-    const { groupBy, report, t } = this.props;
+  private getItems = () => {
+    const { groupBy, report } = this.props;
 
-    const cost = formatCurrency(
-      report && report.meta && report.meta.total
-        ? report.meta.total.cost.value
-        : 0
+    const computedItems = getComputedAwsReportItems({
+      report,
+      idKey: groupBy as any,
+    });
+    return computedItems;
+  };
+
+  private getTabItem = reportItem => {
+    const { report } = this.props;
+
+    return (
+      <AwsReportSummaryItem
+        key={`${reportItem.id}-item`}
+        formatOptions={{}}
+        formatValue={formatValue}
+        label={reportItem.label ? reportItem.label.toString() : ''}
+        totalValue={
+          reportType === AwsReportType.cost
+            ? report.meta.total.cost.value
+            : report.meta.total.usage.value
+        }
+        units={reportItem.units}
+        value={
+          reportType === AwsReportType.cost ? reportItem.cost : reportItem.usage
+        }
+      />
     );
+  };
+
+  private getViewAll = () => {
+    const { groupBy, item, parentGroupBy, t } = this.props;
+    const { isWidgetModalOpen } = this.state;
+    const computedItems = this.getItems();
+
+    const otherIndex = computedItems.findIndex(i => {
+      const id = i.id;
+      if (id && id !== null) {
+        return id.toString().includes('Other');
+      }
+    });
+
+    if (otherIndex !== -1) {
+      return (
+        <div className={css(styles.viewAllContainer)}>
+          <Button
+            {...getTestProps(testIds.details.view_all_btn)}
+            onClick={this.handleWidgetModalOpen}
+            type={ButtonType.button}
+            variant={ButtonVariant.link}
+          >
+            {t('aws_details.view_all', { value: groupBy })}
+          </Button>
+          <DetailsWidgetModal
+            groupBy={groupBy}
+            isOpen={isWidgetModalOpen}
+            item={item}
+            onClose={this.handleWidgetModalClose}
+            parentGroupBy={parentGroupBy}
+          />
+        </div>
+      );
+    } else {
+      return null;
+    }
+  };
+
+  private handleWidgetModalClose = (isOpen: boolean) => {
+    this.setState({ isWidgetModalOpen: isOpen });
+  };
+
+  private handleWidgetModalOpen = event => {
+    this.setState({ isWidgetModalOpen: true });
+    event.preventDefault();
+  };
+
+  public render() {
+    const { groupBy, report, reportFetchStatus } = this.props;
 
     return (
       <>
-        <div className={styles.subTitle}>
-          <Title size="lg">
-            {t('aws_details.cost_value', { value: cost })}
-          </Title>
-        </div>
-        <div className={styles.mainContent}>
-          <AwsReportSummaryItems idKey={groupBy as any} report={report}>
-            {({ items }) =>
-              items.map(_item => (
-                <AwsReportSummaryItem
-                  key={_item.id}
-                  formatOptions={{}}
-                  formatValue={formatValue}
-                  label={_item.label ? _item.label.toString() : ''}
-                  totalValue={report.meta.total.cost.value}
-                  units={_item.units}
-                  value={_item.cost}
-                />
-              ))
-            }
-          </AwsReportSummaryItems>
-        </div>
+        {Boolean(reportFetchStatus === FetchStatus.inProgress) ? (
+          <>
+            <Skeleton size={SkeletonSize.md} />
+            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
+            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
+            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
+          </>
+        ) : (
+          <>
+            <div className={css(styles.tabs)}>
+              <AwsReportSummaryItems
+                idKey={groupBy as any}
+                key={`${groupBy}-items`}
+                report={report}
+              >
+                {({ items }) =>
+                  items.map(reportItem => this.getTabItem(reportItem))
+                }
+              </AwsReportSummaryItems>
+            </div>
+            {this.getViewAll()}
+          </>
+        )}
       </>
     );
   }
@@ -93,12 +183,14 @@ class DetailsWidgetViewBase extends React.Component<DetailsWidgetViewProps> {
 const mapStateToProps = createMapStateToProps<
   DetailsWidgetViewOwnProps,
   DetailsWidgetViewStateProps
->((state, { groupBy }) => {
+>((state, { groupBy, item, parentGroupBy }) => {
   const query: AwsQuery = {
     filter: {
+      limit: 3,
       time_scope_units: 'month',
       time_scope_value: -1,
       resolution: 'monthly',
+      [parentGroupBy]: item.label || item.id,
     },
     group_by: { [groupBy]: '*' },
   };
