+++ src/pages/ocpOnAwsDetails/detailsTableItem.tsx
@@ -12,7 +12,6 @@ import React from 'react';
 import { InjectedTranslateProps, translate } from 'react-i18next';
 import { connect } from 'react-redux';
 import { createMapStateToProps } from 'store/common';
-import { ocpOnAwsDetailsSelectors } from 'store/ocpOnAwsDetails';
 import { getTestProps, testIds } from 'testIds';
 import { ComputedOcpOnAwsReportItem } from 'utils/getComputedOcpOnAwsReportItems';
 import { DetailsChart } from './detailsChart';
@@ -30,18 +29,7 @@ interface DetailsTableItemState {
   isHistoricalModalOpen: boolean;
 }
 
-interface DetailsTableItemStateProps {
-  widgets: number[];
-}
-
-interface DetailsTableItemDispatchProps {
-  selectWidgets?: typeof ocpOnAwsDetailsSelectors.selectWidgets;
-}
-
-type DetailsTableItemProps = DetailsTableItemOwnProps &
-  DetailsTableItemStateProps &
-  DetailsTableItemDispatchProps &
-  InjectedTranslateProps;
+type DetailsTableItemProps = DetailsTableItemOwnProps & InjectedTranslateProps;
 
 class DetailsTableItemBase extends React.Component<DetailsTableItemProps> {
   public state: DetailsTableItemState = {
@@ -65,7 +53,7 @@ class DetailsTableItemBase extends React.Component<DetailsTableItemProps> {
   };
 
   public render() {
-    const { item, groupBy, t, widgets } = this.props;
+    const { item, groupBy, t } = this.props;
     const { isHistoricalModalOpen } = this.state;
 
     return (
@@ -97,16 +85,7 @@ class DetailsTableItemBase extends React.Component<DetailsTableItemProps> {
                   </Form>
                 </div>
               )}
-              {widgets.map(widgetId => {
-                return (
-                  <DetailsWidget
-                    groupBy={groupBy}
-                    item={item}
-                    key={`details-widget-${widgetId}`}
-                    widgetId={widgetId}
-                  />
-                );
-              })}
+              <DetailsWidget groupBy={groupBy} item={item} />
             </div>
           </GridItem>
           <GridItem lg={12} xl={6}>
@@ -143,14 +122,11 @@ class DetailsTableItemBase extends React.Component<DetailsTableItemProps> {
   }
 }
 
-const mapStateToProps = createMapStateToProps<
-  DetailsTableItemOwnProps,
-  DetailsTableItemStateProps
->(state => {
-  return {
-    widgets: ocpOnAwsDetailsSelectors.selectCurrentWidgets(state),
-  };
-});
+const mapStateToProps = createMapStateToProps<DetailsTableItemOwnProps, {}>(
+  state => {
+    return {};
+  }
+);
 
 const DetailsTableItem = translate()(
   connect(
