+++ src/pages/ocpOnAwsDetails/detailsWidget.tsx
@@ -1,67 +1,25 @@
-import {
-  Button,
-  ButtonType,
-  ButtonVariant,
-  Tab,
-  Tabs,
-} from '@patternfly/react-core';
-import { css } from '@patternfly/react-styles';
-import {
-  Skeleton,
-  SkeletonSize,
-} from '@red-hat-insights/insights-frontend-components/components/Skeleton';
-import { OcpOnAwsReport, OcpOnAwsReportType } from 'api/ocpOnAwsReports';
-import {
-  OcpOnAwsReportSummaryItem,
-  OcpOnAwsReportSummaryItems,
-} from 'components/reports/ocpOnAwsReportSummary';
+import { Tab, Tabs } from '@patternfly/react-core';
 import React from 'react';
 import { InjectedTranslateProps, translate } from 'react-i18next';
 import { connect } from 'react-redux';
-import { createMapStateToProps, FetchStatus } from 'store/common';
-import {
-  ocpOnAwsDetailsActions,
-  ocpOnAwsDetailsSelectors,
-  OcpOnAwsDetailsTab,
-  OcpOnAwsDetailsWidget as OcpOnAwsDetailsWidgetStatic,
-} from 'store/ocpOnAwsDetails';
-import { ocpOnAwsReportsSelectors } from 'store/ocpOnAwsReports';
-import { getTestProps, testIds } from 'testIds';
-import { formatValue } from 'utils/formatValue';
+import { createMapStateToProps } from 'store/common';
 import {
   ComputedOcpOnAwsReportItem,
-  getComputedOcpOnAwsReportItems,
   GetComputedOcpOnAwsReportItemsParams,
 } from 'utils/getComputedOcpOnAwsReportItems';
-import { styles } from './detailsWidget.styles';
-import { DetailsWidgetModal } from './detailsWidgetModal';
+import { DetailsWidgetView } from './detailsWidgetView';
 
 interface DetailsWidgetOwnProps {
+  availableTabs?: OcpOnAwsDetailsTab[];
   groupBy: string;
   item: ComputedOcpOnAwsReportItem;
-  widgetId: number;
-}
-
-interface DetailsWidgetStateProps extends OcpOnAwsDetailsWidgetStatic {
-  query: string;
-  report: OcpOnAwsReport;
-  reportFetchStatus: FetchStatus;
 }
 
 interface DetailsWidgetState {
   activeTabKey: number;
-  isWidgetModalOpen: boolean;
 }
 
-interface DetailsWidgetDispatchProps {
-  fetchReports: typeof ocpOnAwsDetailsActions.fetchWidgetReports;
-  updateTab: typeof ocpOnAwsDetailsActions.changeWidgetTab;
-}
-
-type DetailsWidgetProps = DetailsWidgetOwnProps &
-  DetailsWidgetStateProps &
-  DetailsWidgetDispatchProps &
-  InjectedTranslateProps;
+type DetailsWidgetProps = DetailsWidgetOwnProps & InjectedTranslateProps;
 
 export const getIdKeyForTab = (
   tab: OcpOnAwsDetailsTab
@@ -78,36 +36,19 @@ export const getIdKeyForTab = (
   }
 };
 
+export const enum OcpOnAwsDetailsTab {
+  accounts = 'accounts',
+  projects = 'projects',
+  regions = 'regions',
+  services = 'services',
+}
+
 class DetailsWidgetBase extends React.Component<DetailsWidgetProps> {
   public state: DetailsWidgetState = {
     activeTabKey: 0,
-    isWidgetModalOpen: false,
-  };
-
-  public componentDidMount() {
-    const { fetchReports, id, updateTab, widgetId } = this.props;
-    const availableTabs = this.getAvailableTabs();
-    if (availableTabs) {
-      updateTab(id, availableTabs[0]);
-    }
-    fetchReports(widgetId);
-  }
-
-  private handleWidgetModalClose = (isOpen: boolean) => {
-    this.setState({ isWidgetModalOpen: isOpen });
-  };
-
-  private handleWidgetModalOpen = event => {
-    this.setState({ isWidgetModalOpen: true });
-    event.preventDefault();
   };
 
   private handleTabClick = (event, tabIndex) => {
-    const { id, updateTab } = this.props;
-    const availableTabs = this.getAvailableTabs();
-    const tab = availableTabs[tabIndex];
-
-    updateTab(id, tab);
     this.setState({
       activeTabKey: tabIndex,
     });
@@ -130,45 +71,20 @@ class DetailsWidgetBase extends React.Component<DetailsWidgetProps> {
     return tabs;
   };
 
-  private getItems = (currentTab: string) => {
-    const { report } = this.props;
-
-    const computedItems = getComputedOcpOnAwsReportItems({
-      report,
-      idKey: currentTab as any,
-    });
-    return computedItems;
-  };
-
   private getTab = (tab: OcpOnAwsDetailsTab, index: number) => {
-    const { report } = this.props;
-
-    const currentTab = getIdKeyForTab(tab as OcpOnAwsDetailsTab);
-
     return (
       <Tab
         eventKey={index}
         key={`${getIdKeyForTab(tab)}-tab`}
         title={this.getTabTitle(tab)}
       >
-        <div className={css(styles.tabs)}>
-          <OcpOnAwsReportSummaryItems
-            idKey={currentTab}
-            key={`${currentTab}-items`}
-            report={report}
-          >
-            {({ items }) =>
-              items.map(reportItem => this.getTabItem(tab, reportItem))
-            }
-          </OcpOnAwsReportSummaryItems>
-        </div>
-        {this.getViewAll(tab)}
+        {this.getTabItem(tab)}
       </Tab>
     );
   };
 
-  private getTabItem = (tab: OcpOnAwsDetailsTab, reportItem) => {
-    const { reportType, report, topItems } = this.props;
+  private getTabItem = (tab: OcpOnAwsDetailsTab) => {
+    const { groupBy, item } = this.props;
     const { activeTabKey } = this.state;
 
     const availableTabs = this.getAvailableTabs();
@@ -177,22 +93,10 @@ class DetailsWidgetBase extends React.Component<DetailsWidgetProps> {
 
     if (activeTab === currentTab) {
       return (
-        <OcpOnAwsReportSummaryItem
-          key={`${reportItem.id}-item`}
-          formatOptions={topItems.formatOptions}
-          formatValue={formatValue}
-          label={reportItem.label ? reportItem.label.toString() : ''}
-          totalValue={
-            reportType === OcpOnAwsReportType.cost
-              ? report.meta.total.infrastructure_cost.value
-              : report.meta.total.usage.value
-          }
-          units={reportItem.units}
-          value={
-            reportType === OcpOnAwsReportType.cost
-              ? reportItem.infrastructureCost
-              : reportItem.usage
-          }
+        <DetailsWidgetView
+          groupBy={currentTab}
+          item={item}
+          parentGroupBy={groupBy}
         />
       );
     } else {
@@ -225,95 +129,28 @@ class DetailsWidgetBase extends React.Component<DetailsWidgetProps> {
     return t('group_by.top_ocp_on_aws', { groupBy: key });
   };
 
-  private getViewAll = (tab: OcpOnAwsDetailsTab) => {
-    const { item, groupBy, t } = this.props;
-    const { isWidgetModalOpen } = this.state;
-
-    const currentTab = getIdKeyForTab(tab as OcpOnAwsDetailsTab);
-    const computedItems = this.getItems(currentTab);
-
-    const otherIndex = computedItems.findIndex(i => {
-      const id = i.id;
-      if (id && id !== null) {
-        return id.toString().includes('Other');
-      }
-    });
-
-    if (otherIndex !== -1) {
-      return (
-        <div className={css(styles.viewAllContainer)}>
-          <Button
-            {...getTestProps(testIds.details.view_all_btn)}
-            onClick={this.handleWidgetModalOpen}
-            type={ButtonType.button}
-            variant={ButtonVariant.link}
-          >
-            {t('ocp_on_aws_details.view_all', { value: currentTab })}
-          </Button>
-          <DetailsWidgetModal
-            groupBy={currentTab}
-            isOpen={isWidgetModalOpen}
-            item={item}
-            onClose={this.handleWidgetModalClose}
-            parentGroupBy={groupBy}
-          />
-        </div>
-      );
-    } else {
-      return null;
-    }
-  };
-
   public render() {
-    const { reportFetchStatus } = this.props;
-    return (
-      <>
-        {Boolean(reportFetchStatus === FetchStatus.inProgress) ? (
-          <>
-            <Skeleton size={SkeletonSize.md} />
-            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
-            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
-            <Skeleton size={SkeletonSize.md} className={css(styles.skeleton)} />
-          </>
-        ) : (
-          this.getTabs()
-        )}
-      </>
-    );
+    return <>{this.getTabs()}</>;
   }
 }
 
-const mapStateToProps = createMapStateToProps<
-  DetailsWidgetOwnProps,
-  DetailsWidgetStateProps
->((state, { widgetId }) => {
-  const widget = ocpOnAwsDetailsSelectors.selectWidget(state, widgetId);
-  const queries = ocpOnAwsDetailsSelectors.selectWidgetQueries(state, widgetId);
-  return {
-    ...widget,
-    query: queries.tabs,
-    report: ocpOnAwsReportsSelectors.selectReport(
-      state,
-      widget.reportType,
-      queries.tabs
-    ),
-    reportFetchStatus: ocpOnAwsReportsSelectors.selectReportFetchStatus(
-      state,
-      widget.reportType,
-      queries.tabs
-    ),
-  };
-});
-
-const mapDispatchToProps: DetailsWidgetDispatchProps = {
-  fetchReports: ocpOnAwsDetailsActions.fetchWidgetReports,
-  updateTab: ocpOnAwsDetailsActions.changeWidgetTab,
-};
+const mapStateToProps = createMapStateToProps<DetailsWidgetOwnProps, {}>(
+  state => {
+    return {
+      availableTabs: [
+        OcpOnAwsDetailsTab.projects,
+        OcpOnAwsDetailsTab.services,
+        OcpOnAwsDetailsTab.accounts,
+        OcpOnAwsDetailsTab.regions,
+      ],
+    };
+  }
+);
 
 const DetailsWidget = translate()(
   connect(
     mapStateToProps,
-    mapDispatchToProps
+    {}
   )(DetailsWidgetBase)
 );
