--- src/store/awsDetails/index.ts
@@ -1,17 +1,0 @@
-import * as awsDetailsActions from './awsDetailsActions';
-import {
-  awsDetailsStateKey,
-  AwsDetailsTab,
-  AwsDetailsWidget,
-} from './awsDetailsCommon';
-import { awsDetailsReducer } from './awsDetailsReducer';
-import * as awsDetailsSelectors from './awsDetailsSelectors';
-
-export {
-  awsDetailsStateKey,
-  awsDetailsReducer,
-  awsDetailsActions,
-  awsDetailsSelectors,
-  AwsDetailsTab,
-  AwsDetailsWidget,
-};
