--- src/store/ocpOnAwsDetails/index.ts
@@ -1,17 +1,0 @@
-import * as ocpOnAwsDetailsActions from './ocpOnAwsDetailsActions';
-import {
-  ocpOnAwsDetailsStateKey,
-  OcpOnAwsDetailsTab,
-  OcpOnAwsDetailsWidget,
-} from './ocpOnAwsDetailsCommon';
-import { ocpOnAwsDetailsReducer } from './ocpOnAwsDetailsReducer';
-import * as ocpOnAwsDetailsSelectors from './ocpOnAwsDetailsSelectors';
-
-export {
-  ocpOnAwsDetailsStateKey,
-  ocpOnAwsDetailsReducer,
-  ocpOnAwsDetailsActions,
-  ocpOnAwsDetailsSelectors,
-  OcpOnAwsDetailsTab,
-  OcpOnAwsDetailsWidget,
-};
